pm/lib/ilog.c: Fix couple of MISRA.ETYPE.INAPPR.OPERAND.BINOP.2012
The choice if signed return value is a bit unfortunate here. Logarithms
never return negative values if the input value is a positive
integer. With zero the returned value should be an infinite negative
value, and I do not think it is needed. Anyway I do that separately
later if I have time for it.

Fixes: SYSFW-1510

+++ pm/lib/ilog.c
@@ -101,8 +101,8 @@ s32 ilog64(uint64_t _v)
 	s32 ret;
 	s32 m;
 	ret = _v > 0U;
-	m = (_v > 0xFFFFFFFFU) << 5;
-	v = (uint32_t) (_v >> m);
+	m = (_v > 0xFFFFFFFFU) << 5U;
+	v = (uint32_t) (_v >> (u32) m);
 	ret |= m;
 	v |= v >> 1;
 	v |= v >> 2;
